xz: Fix integer type from uint64_t to uint32_t
lzma_options_lzma.dict_size is uint32_t so use it here too.

Fixes: 5f0c5a04388f8334962c70bc37a8c2ff8f605e0a

+++ src/xz/coder.c
@@ -659,7 +659,7 @@ coder_set_compression_settings(void)
 
 		// Original dictionary sizes are used to show how each
 		// filter's dictionary was reduced.
-		uint64_t orig_dict_size;
+		uint32_t orig_dict_size;
 
 		// Index of the LZMA filter in the filters member. We only
 		// adjust this filter's memusage because we don't know how
